pyproject.toml
[project]
name = "tiny_corpus_prep"
version = "0.2.0"
description = "Minimal, fast corpus prep for tiny GPT pretraining with Polars"
readme = "README.md"
requires-python = ">=3.9"
authors = [{name="you", email="you@example.com"}]
dependencies = [
    "polars>=0.19.0",
    "textstat>=0.7.0",
    "pyarrow>=14.0.0",
    "tqdm>=4.65.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
annotators = [
    "google-generativeai>=0.3.0",
    "python-dotenv>=1.0.0",
]
downloaders = [
    "wikiextractor>=3.0.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
